% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Setup.R
\name{AQUA.EL4P.Setup}
\alias{AQUA.EL4P.Setup}
\title{Initialize EL4P Aquatic Ecology Module}
\usage{
AQUA.EL4P.Setup(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite existing methods}
}
\description{
This function initializes methods and fields for the 'EL4P' MODULE of Aquatic Ecology COMPONENT.
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\examples{
AQUA.EL4P.Setup()
}
