% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Setup.R
\name{AQUA.Emerge.Setup}
\alias{AQUA.Emerge.Setup}
\title{Initialize Emerge Aquatic Ecology Module}
\usage{
AQUA.Emerge.Setup(overwrite = TRUE)
}
\arguments{
\item{overwrite}{overwrite existing methods}
}
\description{
This function initializes methods and fields for the 'Emerge' MODULE of Aquatic Ecology COMPONENT.
It modifies \code{\link{AquaticSite}} and \code{\link{Landscape}} classes.
}
\examples{
Aqua.Emerge.Setup()
}
