% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{clear_EggQ_Landscape}
\alias{clear_EggQ_Landscape}
\title{Clear Egg Queue for Landscape}
\usage{
clear_EggQ_Landscape()
}
\description{
Clear the \code{\link[MASHcpp]{EggQ}} for all aquatic sites by calling \code{\link{clear_EggQ_AquaticSite}}
\itemize{
\item This method is bound to \code{Landscape$clear_EggQ}
}
}
