% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{clear_ImagoQ_AquaticSite}
\alias{clear_ImagoQ_AquaticSite}
\title{Clear ImagoQ}
\usage{
clear_ImagoQ_AquaticSite()
}
\description{
Clear the imago queue at this site, an object of class \code{\link[MASHcpp]{ImagoQ}}
\itemize{
\item This method is bound to \code{AquaticSite$clear_ImagoQ}
}
}
