% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{clear_ImagoQ_Landscape}
\alias{clear_ImagoQ_Landscape}
\title{Clear Imago Queue for Landscape}
\usage{
clear_ImagoQ_Landscape()
}
\description{
Clear the \code{\link[MASHcpp]{ImagoQ}} for all aquatic sites by calling \code{\link{clear_ImagoQ_AquaticSite}}
\itemize{
\item This method is bound to \code{Landscape$clear_ImagoQ}
}
}
