% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{close_MaleCSVCon_Tile}
\alias{close_MaleCSVCon_Tile}
\title{Close Male CSV Connection}
\usage{
close_MaleCSVCon_Tile()
}
\description{
Close the text connection used to write population count data to CSV.
\itemize{
\item This method is bound to \code{Tile$close_MaleCSVCon}
}
}
