% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DEBUG.R
\name{DEBUG.MASH.MICRO}
\alias{DEBUG.MASH.MICRO}
\title{Debug Finalizers for Classes}
\usage{
DEBUG.MASH.MICRO(overwrite = TRUE)
}
\description{
Print when objects are garbage collected. For debugging purposes only.
}
