% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Fit.R
\name{EL4P_fit_Landscape}
\alias{EL4P_fit_Landscape}
\title{EL4P Fit a Landscape}
\usage{
EL4P_fit_Landscape(lambda, W = NULL, Wa = 1, Wb = 1, beta = 32,
  mu_E = 0.168, mu_L = 0.168, mu_P = 0.168, mu_M = 0.123, d_E = 1,
  d_L = 14, d_P = 1)
}
\arguments{
\item{lambda}{see \code{\link{EL4P_getLambda}}}

\item{W}{vector of weights on K (if \code{NULL} weights are sampled from gamma distribution)}

\item{Wa}{if \code{W} not provided}

\item{beta}{daily rate of oviposition of eggs (units in 1/day * 1/female mosquito)}

\item{mu_E}{egg stage daily mortality rate}

\item{mu_L}{larvae stage daily mortality rate}

\item{mu_P}{pupae stage daily mortality rate}

\item{mu_M}{adult female daily mortality rate}

\item{d_E}{mean duration of egg stage (days)}

\item{d_L}{mean duration of larvae stage (days)}

\item{d_P}{mean duration of pupae stage (days)}
}
\description{
\itemize{
\item this method is bound to code{Landscape$EL4P_fit}
}
}
