% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Fit.R
\name{EL4P_getLambda}
\alias{EL4P_getLambda}
\title{EL4P Get Lambda}
\usage{
EL4P_getLambda(R0, b = 0.55, c = 0.15, r = 1/200, a = 0.3, g = 1/12,
  EIP = 12)
}
\arguments{
\item{R0}{basic reproductive number (expected number of infected humans produced per infected mosquito when prevalence in humans is 0)}

\item{b}{mosquito to human transmission efficiency}

\item{c}{human to mosquito transmission efficiency}

\item{r}{human recovery rate}

\item{a}{human feeding rate (fQ; feeding rate multiplied by human blood index)}

\item{g}{daily mortality rate (1/lifespan)}

\item{EIP}{length of entomological incubation period}
}
\description{
Given equilibrium parameters and desired value of \deqn{R_{0}}
}
