% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{get_ActivitySpace_Human}
\alias{get_ActivitySpace_Human}
\title{Get \code{\link[MASHmacro]{Human}} ActivitySpace}
\usage{
get_ActivitySpace_Human()
}
\description{
This function is bound to \code{Human$get_ActivitySpace()}
}
