% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{get_aerialSpray}
\alias{get_aerialSpray}
\title{Get aerialSpray}
\usage{
get_aerialSpray()
}
\arguments{
\item{aerialSpray}{Object}
}
\description{
Return the definition of the Aerial Spray
}
