% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_AquaSites_Landscape}
\alias{get_AquaSites_Landscape}
\title{Get Aquatic Habitat Sites}
\usage{
get_AquaSites_Landscape(ixS = NULL)
}
\arguments{
\item{ixS}{integer index; if \code{NULL} return entire list of sites}
}
\description{
Return single \code{\link{AquaticSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$get_AquaSites}
}
}
