% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{get_attractiveSugarBait}
\alias{get_attractiveSugarBait}
\title{Get attractiveSugarBait}
\usage{
get_attractiveSugarBait()
}
\arguments{
\item{attractiveSugarBait}{Object}
}
\description{
Return the ATSB in the sugar site
}
