% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_batch_Mosquito}
\alias{get_batch_Mosquito}
\title{Get batch}
\usage{
get_batch_Mosquito()
}
\description{
Return mosquito egg batch size (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_bmSize}
}
}
