% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_bmSize_Mosquito}
\alias{get_bmSize_Mosquito}
\title{Get bmSize}
\usage{
get_bmSize_Mosquito()
}
\description{
Return mosquito blood meal size (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_bmSize}
}
}
