% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_directory_Tile}
\alias{get_directory_Tile}
\title{Get Directory}
\usage{
get_directory_Tile()
}
\description{
Return the directory (character) containing output for this tile
\itemize{
\item This method is bound to \code{Tile$get_directory}
}
}
