% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-PeriDomestic.R
\name{get_EggQ_periDomestic_AquaticSite}
\alias{get_EggQ_periDomestic_AquaticSite}
\title{Get EggQ}
\usage{
get_EggQ_periDomestic_AquaticSite()
}
\description{
Get mosquito egg batch queue, object of class \code{\link[MASHcpp]{EggQ}}
\itemize{
\item This method is bound to \code{periDomestic_AquaticSite$get_EggQ}
}
}
