% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_EL4P_AquaticSite}
\alias{get_EL4P_AquaticSite}
\title{Get EL4P}
\usage{
get_EL4P_AquaticSite()
}
\description{
Get aquatic populations, object of class \code{\link[MASHcpp]{EL4P}}
\itemize{
\item This method is bound to \code{AquaticSite$get_EL4P}
}
}
