% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_energy_Mosquito}
\alias{get_energy_Mosquito}
\title{Get energy}
\usage{
get_energy_Mosquito()
}
\description{
Return mosquito energy reserves (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_energy} and \code{MosquitoMale$get_energy}
}
}
