% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_FeedingSites_Landscape}
\alias{get_FeedingSites_Landscape}
\title{Get Blood Feeding Sites}
\usage{
get_FeedingSites_Landscape(ixS = NULL)
}
\arguments{
\item{ixS}{integer index; if \code{NULL} return entire list of sites}
}
\description{
Return single \code{\link{FeedingSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$get_FeedingSites}
}
}
