% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_FeedingSitesN_Landscape}
\alias{get_FeedingSitesN_Landscape}
\title{Get Number of Blood Feeding Sites}
\usage{
get_FeedingSitesN_Landscape()
}
\description{
\itemize{
\item This method is bound to \code{Landscape$get_FeedingSitesN}
}
}
