% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_FemalePop_Tile}
\alias{get_FemalePop_Tile}
\title{Get Female Mosquitoes}
\usage{
get_FemalePop_Tile()
}
\description{
Return embedded female mosquito population object \code{\link{MosquitoPopFemale}}
\itemize{
\item This method is bound to \code{Tile$get_FemalePop}
}
}
