% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{get_FemalePopPointer_HumanPop}
\alias{get_FemalePopPointer_HumanPop}
\title{Get \code{\link{MosquitoPopFemale}} Pointer}
\usage{
get_FemalePopPointer_HumanPop()
}
\description{
Return pointer to the female mosquito population on this landscape.
\itemize{
\item This method is bound to \code{HumanPop$get_FemalePopPointer}
}
}
