% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_FemalePopPointer_Mosquito}
\alias{get_FemalePopPointer_Mosquito}
\title{Get \code{\link{MosquitoPopFemale}} Pointer}
\usage{
get_FemalePopPointer_Mosquito()
}
\description{
Return \code{MosquitoPopFemale$self} enclosing this mosquito (if method for \code{\link{MosquitoFemale}}), or in the same \code{\link{Tile}} (if method for \code{\link{MosquitoMale}}) by reference.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_FemalePopPointer} and \code{MosquitoMale$get_FemalePopPointer}
}
}
