% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_FemalePopPointer_MosquitoPopMale}
\alias{get_FemalePopPointer_MosquitoPopMale}
\title{Get \code{\link{MosquitoPopFemale}} Pointer}
\usage{
get_FemalePopPointer_MosquitoPopMale()
}
\description{
Return \code{MosquitoPopFemale$self} in the same \code{\link{Tile}} by reference.
\itemize{
\item This method is bound to \code{MosquitoPopMale$get_FemalePopPointer}
}
}
