% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_haz_MatingSite}
\alias{get_haz_MatingSite}
\title{Get haz}
\usage{
get_haz_MatingSite()
}
\description{
Get landing hazard associated with outside wall at this site \code{haz}
\itemize{
\item This method is bound to \code{MatingSite$get_haz}
}
}
