% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_haz_SugarSite}
\alias{get_haz_SugarSite}
\title{Get haz}
\usage{
get_haz_SugarSite()
}
\description{
Get landing hazard associated with outside wall at this site \code{haz}
\itemize{
\item This method is bound to \code{SugarSite$get_haz}
}
}
