% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_hazI_FeedingSite}
\alias{get_hazI_FeedingSite}
\title{Get hazI}
\usage{
get_hazI_FeedingSite()
}
\description{
Get landing hazard associated with indoor wall at this site \code{hazI}
\itemize{
\item This method is bound to \code{FeedingSite$get_hazI}
}
}
