% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_HumanDirectory_Tile}
\alias{get_HumanDirectory_Tile}
\title{Get Directory for Human Output}
\usage{
get_HumanDirectory_Tile()
}
\description{
Return the directory (character) containing human output for this tile
\itemize{
\item This method is bound to \code{Tile$get_HumanDirectory}
}
}
