% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_HumanPathogenCon_Tile}
\alias{get_HumanPathogenCon_Tile}
\title{Get Human Pathogen JSON Connection}
\usage{
get_HumanPathogenCon_Tile()
}
\description{
Return the text connection used to write human-stage pathogen data to JSON
\itemize{
\item This method is bound to \code{Tile$get_HumanPathogenCon}
}
}
