% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_HumansPointer_Landscape}
\alias{get_HumansPointer_Landscape}
\title{Get \code{\link[MASHmacro]{HumanPop}} Pointer}
\usage{
get_HumansPointer_Landscape()
}
\description{
Return pointer to \code{HumanPop$self}.
\itemize{
\item This method is bound to \code{Landscape$get_HumansPointer}
}
}
