% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_inPointSet_Mosquito}
\alias{get_inPointSet_Mosquito}
\title{Get inPointSet}
\usage{
get_inPointSet_Mosquito()
}
\description{
Return mosquito point set (\code{character}). With \code{ix}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_inPointSet} and \code{MosquitoMale$get_inPointSet}
}
}
