% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_ix_FeedingSite}
\alias{get_ix_FeedingSite}
\title{Get ix}
\usage{
get_ix_FeedingSite()
}
\description{
Return site index \code{ix}
\itemize{
\item This method is bound to \code{FeedingSite$get_ix}
}
}
