% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_ix_Mosquito}
\alias{get_ix_Mosquito}
\title{Get ix}
\usage{
get_ix_Mosquito()
}
\description{
Return mosquito point index (\code{integer}). With \code{inPointSet}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_ix} and \code{MosquitoMale$get_ix}
}
}
