% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_Landscape_Tile}
\alias{get_Landscape_Tile}
\title{Get Landscape}
\usage{
get_Landscape_Tile()
}
\description{
Return embedded landscape object \code{\link{Landscape}}
\itemize{
\item This method is bound to \code{Tile$get_Landscape}
}
}
