% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_LandscapePointer_Mosquito}
\alias{get_LandscapePointer_Mosquito}
\title{Get \code{\link{Landscape}} Pointer}
\usage{
get_LandscapePointer_Mosquito()
}
\description{
Return enclosing \code{Landscape$self} by refernce.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_LandscapePointer} and \code{MosquitoMale$get_LandscapePointer}
}
}
