% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_LandscapePointer_MosquitoPop}
\alias{get_LandscapePointer_MosquitoPop}
\title{Get \code{\link{Landscape}} Pointer}
\usage{
get_LandscapePointer_MosquitoPop()
}
\description{
Return enclosing \code{Landscape$self} by refernce.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$get_LandscapePointer} and \code{MosquitoPopMale$get_LandscapePointer}
}
}
