% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{get_larviciding}
\alias{get_larviciding}
\title{Get swarmSpray}
\usage{
get_larviciding()
}
\arguments{
\item{swarmSpray}{Object}
}
\description{
Return the definition of the Swarm Spray
}
