% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_locNext_Mosquito}
\alias{get_locNext_Mosquito}
\title{Get locNext}
\usage{
get_locNext_Mosquito()
}
\description{
Return next mosquito point set (\code{character}). With \code{pSetNext}, the next site of the mosquito's spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_locNext} and \code{MosquitoMale$get_locNext}
}
}
