% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_MalePop_Tile}
\alias{get_MalePop_Tile}
\title{Get Male Mosquitoes}
\usage{
get_MalePop_Tile()
}
\description{
Return embedded male mosquito population object \code{\link{MosquitoPopMale}}
\itemize{
\item This method is bound to \code{Tile$get_MalePop}
}
}
