% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{get_MalePopPointer_Human}
\alias{get_MalePopPointer_Human}
\title{Get \code{\link{MosquitoPopMale}} Pointer}
\usage{
get_MalePopPointer_Human()
}
\description{
Return pointer to the female mosquito population on this landscape.
\itemize{
\item This method is bound to \code{Human$get_MalePopPointer}
}
}
