% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_MalePopPointer_Landscape}
\alias{get_MalePopPointer_Landscape}
\title{Get \code{\link{MosquitoPopMale}} Pointer}
\usage{
get_MalePopPointer_Landscape()
}
\description{
Return pointer to \code{MosquitoPopMale$self}.
\itemize{
\item This method is bound to \code{Landscape$get_MalePopPointer}
}
}
