% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{get_MalePopPointer_MosquitoPopFemale}
\alias{get_MalePopPointer_MosquitoPopFemale}
\title{Get \code{\link{MosquitoPopMale}} Pointer}
\usage{
get_MalePopPointer_MosquitoPopFemale()
}
\description{
Return \code{MosquitoPopMale$self} in the same \code{\link{Tile}} by reference.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$get_MalePopPointer}
}
}
