% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_MatingSitesN_Landscape}
\alias{get_MatingSitesN_Landscape}
\title{Get Number of Mating Sites}
\usage{
get_MatingSitesN_Landscape()
}
\description{
\itemize{
\item This method is bound to \code{Landscape$get_MatingSitesN}
}
}
