% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_mature_Mosquito}
\alias{get_mature_Mosquito}
\title{Get mature}
\usage{
get_mature_Mosquito()
}
\description{
Return mosquito maturity (\code{logical}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_mature} and \code{MosquitoMale$get_mature}
}
}
