% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Methods.R
\name{get_MicroKernel_movement_Male}
\alias{get_MicroKernel_movement_Male}
\title{MICRO Search Kernels: \code{\link{MicroTile}} Access MvAll Object for Male M-BITES Lifecycle Models}
\usage{
get_MicroKernel_movement_Male(locNow, state, pSetNow)
}
\description{
Replace generic \code{MicroTile$get_movement()} method for MicroKernel module, return a MvOb from \code{movementMale} field of a microsimulation tile.
\itemize{
\item This method bound to \code{MicroTile$get_movementMale}
}
}
