% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_MosquitoDirectory_Tile}
\alias{get_MosquitoDirectory_Tile}
\title{Get Directory for Mosquito Output}
\usage{
get_MosquitoDirectory_Tile()
}
\description{
Return the directory (character) containing mosquito output for this tile
\itemize{
\item This method is bound to \code{Tile$get_MosquitoDirectory}
}
}
