% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_movementFemale_Tile}
\alias{get_movementFemale_Tile}
\title{Get movementFemale}
\usage{
get_movementFemale_Tile()
}
\description{
Return female movement object
\itemize{
\item This method is bound to \code{Tile$get_movementFemale}
}
}
