% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{get_movementMale_Tile}
\alias{get_movementMale_Tile}
\title{Get movementMale}
\usage{
get_movementMale_Tile()
}
\description{
Return female movement object
\itemize{
\item This method is bound to \code{Tile$get_movementMale}
}
}
