% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_numGenotypes_AquaticSite}
\alias{get_numGenotypes_AquaticSite}
\title{Get numGenotypes}
\usage{
get_numGenotypes_AquaticSite()
}
\description{
Get number of genotypes.
\itemize{
\item This method is bound to \code{AquaticSite$get_numGenotypes}
}
}
