% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_periDomestic_FeedingSite}
\alias{get_periDomestic_FeedingSite}
\title{Get peri-domestic breeding sites}
\usage{
get_periDomestic_FeedingSite()
}
\description{
Return peri-domestic site attached to this \code{\link{FeedingSite}}, if it exists.
\itemize{
\item This method is bound to \code{FeedingSite$get_periDomestic}
}
}
