% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_PointSet_Landscape}
\alias{get_PointSet_Landscape}
\title{Get Point Set of Sites}
\usage{
get_PointSet_Landscape()
}
\value{
list
}
\description{
Return point set of sites and associated search weights of a landscape.
\itemize{
\item This method is bound to \code{Landscape$get_PointSet}
}
}
