% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_pSetNext_Mosquito}
\alias{get_pSetNext_Mosquito}
\title{Get pSetNext}
\usage{
get_pSetNext_Mosquito()
}
\description{
Return next mosquito point set (\code{character}). With \code{locNext}, the next site of the mosquito's spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_pSetNext} and \code{MosquitoMale$get_pSetNext}
}
}
