% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_pSetNow_Mosquito}
\alias{get_pSetNow_Mosquito}
\title{Get pSetNow}
\usage{
get_pSetNow_Mosquito()
}
\description{
Return current mosquito point set (\code{character}). With \code{locNow}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_pSetNow} and \code{MosquitoMale$get_pSetNow}
}
}
