% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_searchWt_SugarSite}
\alias{get_searchWt_SugarSite}
\title{Get searchWt}
\usage{
get_searchWt_SugarSite()
}
\description{
Get site search weight \code{searchWt}
\itemize{
\item This method is bound to \code{SugarSite$get_searchWt}
}
}
