% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_sire_Mosquito}
\alias{get_sire_Mosquito}
\title{Get sire}
\usage{
get_sire_Mosquito()
}
\description{
Return ID of the \code{\link{MosquitoMale}} this mosquito mated with (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_sire}
}
}
