% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{get_siteType_MatingSite}
\alias{get_siteType_MatingSite}
\title{Get siteType}
\usage{
get_siteType_MatingSite()
}
\description{
Get site type \code{siteType}; 1 is peri-domestic, 0 is non peri-domestic.
\itemize{
\item This method is bound to \code{MatingSite$get_siteType}
}
}
