% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{get_TilePointer_Landscape}
\alias{get_TilePointer_Landscape}
\title{Get \code{\link{Tile}} Pointer}
\usage{
get_TilePointer_Landscape()
}
\description{
Return pointer to enclosing \code{Tile$self}.
\itemize{
\item This method is bound to \code{Landscape$get_TilePointer}
}
}
