% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{get_tNow_Mosquito}
\alias{get_tNow_Mosquito}
\title{Get tNow}
\usage{
get_tNow_Mosquito()
}
\description{
Return mosquito current time (time at which mosquito entered current state) (\code{numeric}).
\itemize{
\item This method is bound to \code{MosquitoFemale$get_tNow} and \code{MosquitoMale$get_tNow}
}
}
