% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{Humans.Setup}
\alias{Humans.Setup}
\title{Initialize Human Methods for MICRO}
\usage{
Humans.Setup(overwrite = TRUE)
}
\description{
Initialize additional methods in \code{\link[MASHmacro]{Human}} and \code{\link[MASHmacro]{HumanPop}}
}
