% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Parameters.R
\name{MBITES.Complex.Parameters}
\alias{MBITES.Complex.Parameters}
\title{Generate Parameters for M-BITES Complex Module}
\usage{
MBITES.Complex.Parameters(SUGAR = TRUE, MATE = TRUE, F_time = 1,
  B_time = 0.75, R_time = 1.5, L_time = 0.75, O_time = 1,
  M_time = 1.5, S_time = 0.5, gammaShape = 8, F_succeed = 0.99,
  B_succeed = 0.99, L_succeed = 0.99, O_succeed = 0.99,
  M_succeed = 0.95, S_succeed = 0.95, F_surv = 0.95, B_surv = 0.98,
  R_surv = 0.98, L_surv = 0.8, O_surv = 0.98, M_surv = 0.98,
  S_surv = 0.98, F_wts = rep(1, 5), B_wts = rep(1, 5), R_wts = rep(1,
  5), L_wts = rep(1, 5), O_wts = rep(1, 5), M_wts = rep(1, 5),
  S_wts = rep(1, 5), InAndOut = matrix(data = c(4, 2, 2, 1, 6, 2, 1, 1, 1,
  4, 1, 1, 1, 1, 2, 0, 0, 0, 0, 1, 1, 1, 2, 1, 0), nrow = 5, ncol = 5, byrow =
  TRUE, dimnames = list(c("i", "w", "v", "r", "l"), c("i", "w", "v", "r",
  "l"))), surviveH = 1, probeH = 1, surviveprobeH = 1, feedH = 1,
  surviveZ = 1, feedZ = 1, bm.a = 7.5, bm.b = 2.5, REFEED = FALSE,
  rf.a = 60, rf.b = 5000, OVERFEED = FALSE, of.a = 5, of.b = 5000,
  SENESCE = FALSE, sns.a = 0.085, sns.b = 100, TATTER = FALSE,
  ttsz.p = 0.5, ttsz.a = 5, ttsz.b = 95, ttr.a = 15, ttr.b = 500,
  S.u = 1/7, S.a = 20, S.b = 10, S.sa = 15, S.sb = 5, bs.m = 30,
  bs.v = 5, maxBatch = 30, emt.m = 3, emt.v = 1, eggT = 0, eggP = 0,
  energyPreG = 0, preGsugar = 0, preGblood = 0, PfEIP = 12, Q = 0.9)
}
\arguments{
\item{F_time}{mean time elapsed (in days) during blood feeding search bout}

\item{B_time}{mean time elapsed (in days) during blood feeding attempt bout}

\item{R_time}{mean time elapsed (in days) during post-prandial resting bout}

\item{L_time}{mean time elapsed (in days) during egg laying search bout}

\item{O_time}{mean time elapsed (in days) during egg laying attempt bout}

\item{M_time}{mean time elapsed (in days) during mating bout}

\item{S_time}{mean time elapsed (in days) during sugar feeding attempt bout}

\item{gammaShape}{shape parameter for Gamma distributed dwell times; if gammaShape = 1 same as exponential, variance decreases as gammaShape becomes large (see \code{\link{mbites_timingGamma}})}

\item{F_succeed}{blood feeding search bout probability of success}

\item{B_succeed}{blood feeding attempt bout probability of success}

\item{L_succeed}{egg laying search bout probability of success}

\item{O_succeed}{egg laying attempt bout probability of success}

\item{M_succeed}{mating bout probability of success}

\item{S_succeed}{sugar feeding attempt bout probability of success}

\item{F_surv}{blood feeding search bout baseline survival probaility (see \code{\link{mbites_surviveFlight}})}

\item{B_surv}{blood feeding attempt bout baseline survival probaility (see \code{\link{mbites_surviveFlight}})}

\item{R_surv}{post-prandial resting bout baseline survival probaility (see \code{\link{mbites_surviveFlight}})}

\item{L_surv}{egg laying search bout baseline survival probaility (see \code{\link{mbites_surviveFlight}})}

\item{O_surv}{egg laying attempt bout baseline survival probaility (see \code{\link{mbites_surviveFlight}})}

\item{M_surv}{mating bout baseline survival probaility (see \code{\link{mbites_surviveFlight}})}

\item{S_surv}{sugar feeding attempt bout baseline survival probaility (see \code{\link{mbites_surviveFlight}})}

\item{F_wts}{blood feeding search bout landing spot weights (see \code{\link{mbites_get_WTS}})}

\item{B_wts}{blood feeding attempt bout landing spot weights (see \code{\link{mbites_get_WTS}})}

\item{R_wts}{post-prandial resting bout landing spot weights (see \code{\link{mbites_get_WTS}})}

\item{L_wts}{egg laying search bout landing spot weights (see \code{\link{mbites_get_WTS}})}

\item{O_wts}{egg laying attempt bout landing spot weights (see \code{\link{mbites_get_WTS}})}

\item{M_wts}{mating bout landing spot weights (see \code{\link{mbites_get_WTS}})}

\item{S_wts}{sugar feeding attempt bout landing spot weights (see \code{\link{mbites_get_WTS}})}

\item{InAndOut}{unnormalized transition matrix of landing probabilities}

\item{surviveH}{human host encounter survival probability of initial encounter (proceed to probe)}

\item{probeH}{human host encounter probability undeterred and begin probing}

\item{surviveprobeH}{human host encounter survival probability of probing}

\item{feedH}{human host encounter probability to successfully blood feed}

\item{surviveZ}{non-human host encounter survival probability of initial encounter (proceed to feed)}

\item{feedZ}{non-human host encounter probability to successfully blood feed}

\item{bm.a}{shape param for bloodmeal size (see \code{\link{mbites_rBloodMealSize}})}

\item{bm.b}{shape param for bloodmeal size (see \code{\link{mbites_rBloodMealSize}})}

\item{REFEED}{control boolean for refeeding behavior during resting bout}

\item{rf.a}{exp param for refeeding as function of bmSize (see \code{\link{mbites_pReFeed}})}

\item{rf.b}{exp param for refeeding as function of bmSize (see \code{\link{mbites_pReFeed}})}

\item{OVERFEED}{control boolean for overfeeding during \code{\link{mbites_BloodMeal}}}

\item{of.a}{exp param for overfeeding as function of bmSize (see \code{\link{mbites_pOverFeed}})}

\item{of.b}{exp param for overfeeding as function of bmSize (see \code{\link{mbites_pOverFeed}})}

\item{SENESCE}{control boolean for senesce mortality during \code{\link{mbites_surviveFlight}}}

\item{sns.a}{exp param for senescence (see \code{\link{mbites_pSenesce}})}

\item{sns.b}{exp param for senescence (see \code{\link{mbites_pSenesce}})}

\item{TATTER}{control boolean for wing tattering during \code{\link{mbites_surviveFlight}}}

\item{ttsz.p}{zero-inflation for tattering damage (see \code{\link{mbites_WingTattering}})}

\item{ttsz.a}{shape param for tattering damage (see \code{\link{mbites_WingTattering}})}

\item{ttsz.b}{shape param for tattering damage (see \code{\link{mbites_WingTattering}})}

\item{ttr.a}{exp param for tattering survival (see \code{\link{mbites_pTatter}})}

\item{ttr.b}{exp param for tattering survival (see \code{\link{mbites_pTatter}})}

\item{S.u}{per-bout energy expenditure}

\item{S.a}{shape parameter of per-bout probability of survival as function of energy reserves}

\item{S.b}{shape parameter of per-bout probability of survival as function of energy reserves}

\item{S.sa}{shape parameter of probability to queue sugar bout as function of energy reserves}

\item{S.sb}{shape parameter of probability to queue sugar bout as function of energy reserves}

\item{bs.m}{mean of normally-distributed egg batch size (used in \code{\link{mbites_rBatchSizeNorm}})}

\item{bs.v}{standard deviation of normally-distributed egg batch size (used in \code{\link{mbites_rBatchSizeNorm}})}

\item{maxBatch}{maximum egg batch size (used in \code{\link{mbites_rBatchSizeBms}})}

\item{emt.m}{mean of normally-distributed egg batch maturation time (used in \code{\link{mbites_rEggMaturationTimeNorm}})}

\item{emt.v}{standard deviation of normally-distributed egg batch maturation time (used in \code{\link{mbites_rEggMaturationTimeNorm}})}

\item{eggT}{minimum time to egg maturation}

\item{eggP}{minimum provision to produce eggs}

\item{energyPreG}{pre-gonotrophic energy requirement}

\item{preGsugar}{amount of energy a sugar meal contributes to energyPreG (pre-gonotrophic energy requirement)}

\item{preGblood}{amount of energy a blood meal contributes to energyPreG (pre-gonotrophic energy requirement)}

\item{PfEIP}{entomological inoculation period for Plasmodium falciparum during \code{MosquitoFemale$probing()}}

\item{Q}{human blood index (used in \code{\link{mbitesCohort_chooseHost}})}
}
\value{
a named list of parameters
}
\description{
Generate named list of parameters used throughout MICRO/M-BITES. All arguments have default values which are listed below before the definition.
}
\examples{
MBITES.Complex.Parameters()
}
