% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Setup.R, R/MBITES-Setup.R
\name{MBITES.Generic.Setup}
\alias{MBITES.Generic.Setup}
\alias{MBITES.Generic.Setup}
\title{Initialize Male Methods for M-BITES}
\usage{
MBITES.Generic.Setup(overwrite = TRUE, batchSize = "bms", eggMatT = "off")

MBITES.Generic.Setup(overwrite = TRUE, batchSize = "bms", eggMatT = "off")
}
\arguments{
\item{batchSize}{character switch that should be one of \code{"bms","norm"} for egg batch sizes dependent on bloodmeal size or normally distributed}

\item{eggMatT}{character switch that should be one of \code{"off","norm"} for egg batch maturation time turned off or normally distributed}

\item{timing}{character in "exponential", "gamma", "weibull"}
}
\value{
writes methods to \code{\link{MosquitoMale}} class.

writes methods to \code{\link{MosquitoFemale}} and \code{\link{MosquitoMale}} classes.
}
\description{
This function initializes methods for M-BITES Male module; this function adds methods to the \code{\link{MosquitoMale}} class.

This function initializes generic methods for M-BITES models; please note that the
switches for this function modify only the methods that are added to the \code{\link{MosquitoFemale}}
and \code{\link{MosquitoMale}} classes.
}
