% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Parameters.R
\name{MBITES.Male.Parameters}
\alias{MBITES.Male.Parameters}
\title{Generate Parameters for M-BITES Male Lifecycle Module}
\usage{
MBITES.Male.Parameters(maleHistory = FALSE, M_wts = rep(1, 5),
  S_wts = rep(1, 5), R_wts = rep(1, 5), InAndOut = matrix(data = c(4, 2,
  2, 1, 6, 2, 1, 1, 1, 4, 1, 1, 1, 1, 2, 0, 0, 0, 0, 1, 1, 1, 2, 1, 0), nrow =
  5, ncol = 5, byrow = TRUE, dimnames = list(c("i", "w", "v", "r", "l"), c("i",
  "w", "v", "r", "l"))), M_surv = 0.85, S_surv = 0.8, R_surv = 0.95,
  M_succeed = 0.95, S_succeed = 0.95, R_succeed = 0.99, gammaShape = 8,
  M_time = 0.5, S_time = 0.65, R_time = 0.75, TATTER = FALSE,
  ttsz.p = 0.5, ttsz.a = 5, ttsz.b = 95, ttr.a = 15, ttr.b = 500,
  SENESCE = FALSE, sns.a = 0.1, sns.b = 100, S.u = 1/5, S.a = 20,
  S.b = 10, S.sa = 15, S.sb = 5)
}
\arguments{
\item{maleHistory}{logical; write male histories to JSON}

\item{M_wts}{landing spot weights}

\item{S_wts}{landing spot weights}

\item{R_wts}{landing spot weights}

\item{InAndOut}{matrix of landing probabilities}

\item{M_surv}{baseline probability to survive flights associated with mating}

\item{S_surv}{baseline probability to survive flights associated with sugar feeding}

\item{R_surv}{baseline probability to survive flights associated with resting}

\item{M_succeed}{probability to successfully enter swarm in mating bout}

\item{S_succeed}{probability to successfully locate sugar source and replenish energy in sugar feeding bout}

\item{R_succeed}{probability to successfully locate suitable resting spot in resting bout}

\item{gammaShape}{shape parameter for Gamma distributed dwell times; if gammaShape = 1 same as exponential, variance decreases as gammaShape becomes large (see \code{\link{mbitesMale_timingGamma}})}

\item{M_time}{average duration of time spent in mating state (unconditional on next state)}

\item{S_time}{average duration of time spent in sugar feeding state (unconditional on next state)}

\item{R_time}{average duration of time spent in resting state (unconditional on next state)}

\item{TATTER}{logical; simulate multiplicative effect on mortality of wing tattering}

\item{ttsz.p}{zero-inflation of Beta distributed per-bout wing tattering damage}

\item{ttsz.a}{alpha parameter of Beta distributed per-bout wing tattering damage}

\item{ttsz.b}{beta parameter of Beta distributed per-bout wing tattering damage}

\item{ttr.a}{shape parameter of per-bout multiplicative effect on mortality of wing tattering damage}

\item{SENESCE}{logical; simulate multiplicative effect on mortality of senescence}

\item{sns.a}{shape parameter of per-bout probability of death due to mosquito age}

\item{sns.b}{shape parameter of per-bout probability of death due to mosquito age}

\item{S.u}{per-bout energy expenditure}

\item{S.a}{shape parameter of per-bout probability of survival as function of energy reserves}

\item{S.b}{shape parameter of per-bout probability of survival as function of energy reserves}

\item{S.sa}{shape parameter of probability to queue sugar bout as function of energy reserves}

\item{S.sb}{shape parameter of probability to queue sugar bout as function of energy reserves}

\item{ttr.a}{shape parameter of per-bout multiplicative effect on mortality of wing tattering damage}
}
\value{
a named list of parameters
}
\description{
Generate named list of parameters used throughout MICRO/M-BITES. All arguments have default values which are listed below before the definition.
}
\examples{
MBITES.Male.Parameters()
}
