% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Search-Setup.R
\name{MBITES.Search.Setup}
\alias{MBITES.Search.Setup}
\title{MICRO Search Kernels: Initialize Additional Methods & Fields in \code{\link{MosquitoPopFemale}} and \code{\link{MosquitoFemale}}}
\usage{
MBITES.Search.Setup(module = "kernel", overwrite = TRUE)
}
\arguments{
\item{module}{character in "kernel"}

\item{overwrite}{overwrite existing methods}
}
\description{
Initialize MICRO Search algorithms for mosquito movement.
}
