% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Setup.R
\name{MBITES.Setup}
\alias{MBITES.Setup}
\title{M-BITES: Initialize Additional Methods & Fields in \code{\link{MosquitoPopFemale}} and \code{\link{MosquitoFemale}}}
\usage{
MBITES.Setup(overwrite = TRUE, SUGAR = TRUE, MATE = TRUE,
  aquaModule = "emerge", timing = "exponential", tattering = "mean",
  energy = "mean", batchSize = "bms", eggMatT = "off")
}
\arguments{
\item{overwrite}{overwrite existing methods}

\item{SUGAR}{enable/disable sugar feeding and sugar energetics}

\item{MATE}{enable/disable mating behavior}

\item{aquaModule}{character switch in \code{"emerge","EL4P"}, selects aquatic ecology module}

\item{timing}{character switch in \code{"exponential","gamma"}, selects waiting time distribution from \code{\link{mbites_timingExponential}} or \code{\link{mbites_timingGamma}}}

\item{tattering}{character switch in \code{"mean","exact"}, mean will sample wing tattering from \code{\link{mbites_WingTattering}} regardless of flight distance}

\item{energy}{character switch in \code{"mean","exact"}, mean will sample energy consumption}

\item{batchSize}{character switch in \code{"bms","norm"} for egg batch sizes dependent on bloodmeal size or normally distributed}

\item{eggMatT}{character switch in \code{"off","norm"} for egg batch maturation time turned off or normally distributed}
}
\value{
writes methods to \code{\link{MosquitoFemale}} and \code{\link{MosquitoMale}} classes.
}
\description{
Initialize M-BITES lifecycle model (complex form). This model includes:
\itemize{
\item Blood Feeding Search Bout: see \code{\link{mbites_boutF}}
\item Blood Feeding Attempt Bout: see \code{\link{mbites_boutB}}
\item Post-Prandial Resting Bout: see \code{\link{mbites_boutR}}
\item Egg Laying Search Bout: see \code{\link{mbites_boutL}}
\item Egg Laying Attempt Bout: see \code{\link{mbites_boutO}}
\item Sugar Feeding Attempt Bout: see \code{\link{mbites_boutS}}
\item Mating Bout: see \code{\link{mbites_boutM}}
}
}
