% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_boutM}
\alias{mbites_boutM}
\title{M-BITES: Mating Bout \code{MosquitoFemale}}
\usage{
mbites_boutM()
}
\description{
A mosquito performs mating bout (all actions taken launch to launch when resting required).
Upon entering the mating behavioral state, the mosquito will move to a \code{\link{MatingSite}} and
finds a mate in the \code{\link[MASHcpp]{MatingQ}}.
\itemize{
\item This method is bound to \code{MosquitoFemale$boutR()}.
}
}
