% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_boutS}
\alias{mbites_boutS}
\title{M-BITES: Sugar Feeding Bout (S) \code{MosquitoFemale}}
\usage{
mbites_boutS()
}
\description{
A mosquito performs a sugar feeding bout (all actions taken launch to launch when resting required).
Upon entering the sugar feeding behavioral state prompted by \code{\link{mbites_queueSugarBout}}, the mosquito will move to a \code{\link{SugarSite}} and
sugar feed .
\itemize{
\item This method is bound to \code{MosquitoFemale$boutS()}.
}
}
