% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-ChooseHost.R
\name{mbites_chooseHost}
\alias{mbites_chooseHost}
\title{MBITES-Generic: Host Choosing for \code{\link{MosquitoFemale}}}
\usage{
mbites_chooseHost()
}
\description{
Choose a human or animal host at a site.
\itemize{
\item This method is bound to \code{MosquitoFemale$chooseHost()}.
}
}
