% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-ChooseMate.R
\name{mbites_chooseMate}
\alias{mbites_chooseMate}
\title{MBITES-Generic: Mate Choosing for \code{\link{MosquitoFemale}}}
\usage{
mbites_chooseMate()
}
\description{
Choose a \code{\link{MosquitoMale}}
\itemize{
\item This method is bound to \code{MosquitoFemale$chooseMate()}.
}
}
