% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_enterHouse}
\alias{mbites_enterHouse}
\title{M-BITES: Attempt to Enter a House for \code{MosquitoFemale}}
\usage{
mbites_enterHouse()
}
\description{
Method to simulate attempted house entry for mosquito, and call appropriate events if the mosquito enters.
\itemize{
\item This method is bound to \code{MosquitoFemale$enterHouse()}.
}
}
