% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_flightEnergetics_Exact}
\alias{mbites_flightEnergetics_Exact}
\title{MBITES-Generic: Exact Flight Energetics for \code{\link{MosquitoFemale}}}
\usage{
mbites_flightEnergetics_Exact()
}
\description{
Reduce this mosquito's energy reserves by Beta-distributed random variable as a function of flight distance.
\itemize{
\item This method is bound to \code{MosquitoFemale$flightEnergetics()}.
}
}
