% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_flightEnergetics_Mean}
\alias{mbites_flightEnergetics_Mean}
\title{MBITES-Generic: Mean Flight Energetics for \code{\link{MosquitoFemale}}}
\usage{
mbites_flightEnergetics_Mean()
}
\description{
Reduce this mosquito's energy reserves by mean amount averaged over possible flight distances.
\itemize{
\item This method is bound to \code{MosquitoFemale$flightEnergetics()}.
}
}
