% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_get_restHaz}
\alias{mbites_get_restHaz}
\title{M-BITES: Get Resting Hazards for \code{\link{MosquitoFemale}}}
\usage{
mbites_get_restHaz()
}
\description{
Get resting hazards for \code{\link{mbites_surviveResting}}.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_restHaz()}.
}
}
