% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_get_WTS}
\alias{mbites_get_WTS}
\title{M-BITES: Return Landing Spot Weights for \code{MosquitoFemale}}
\usage{
mbites_get_WTS()
}
\value{
vector of landing spot weights
}
\description{
Method for return landing spot weights based on behavioral state of mosquito.
\itemize{
\item This method is bound to \code{MosquitoFemale$get_WTS()}.
}
}
