% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-HostEncounter.R
\name{mbites_humanEncounter}
\alias{mbites_humanEncounter}
\title{MBITES-BRO: Human Host Encounter for \code{\link{MosquitoFemale}}}
\usage{
mbites_humanEncounter()
}
\description{
After calling \code{\link{mbites_chooseHost}}, the mosquito encounters the human host and attempts to feed.
\itemize{
\item This method is bound to \code{MosquitoFemale$humanEncounter()}.
}
}
