% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Auxiliary.R
\name{mbites_isActive}
\alias{mbites_isActive}
\title{MBITES-Generic: Active Check for \code{\link{MosquitoFemale}} and \code{\link{MosquitoMale}}}
\usage{
mbites_isActive()
}
\description{
Check if this mosquito is active and return a logical value.
\itemize{
\item This method is bound to \code{MosquitoFemale$isActive()} or \code{MosquitoMale$isActive()}.
}
}
