% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_layEggs_EL4P}
\alias{mbites_layEggs_EL4P}
\title{M-BITES: Lay Eggs for 'EL4P' \code{\link{MosquitoFemale}}}
\usage{
mbites_layEggs_EL4P()
}
\description{
During an oviposition bout \code{\link{mbites_boutO}}, lay eggs for 'EL4P' module of Aquatic Ecology.
\itemize{
\item This method is bound to \code{MosquitoFemale$layEggs()}.
}
}
