% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Complex-Bouts.R
\name{mbites_layEggs_Emerge}
\alias{mbites_layEggs_Emerge}
\title{M-BITES: Lay Eggs for 'Emerge' \code{\link{MosquitoFemale}}}
\usage{
mbites_layEggs_Emerge()
}
\description{
During an oviposition bout \code{\link{mbites_boutO}}, lay eggs (this is just a filler to clear out the \code{batch} field of the mosquito; egg laying is not implemented in any modules relying on "Emerge" Aquatic Ecology module)
\itemize{
\item This method is bound to \code{MosquitoFemale$layEggs()}.
}
}
