% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_newSpot}
\alias{mbites_newSpot}
\title{M-BITES: Generate New Landing Spot for \code{MosquitoFemale}}
\usage{
mbites_newSpot()
}
\value{
integer value corresponding to new landing spot
}
\description{
Method for return a new landing spot based on behavioral state of mosquito and weights from \code{\link{mbites_get_WTS}}.
New landing spots generated at the end of the search bout, attempt bout, or after oviposition a mosquito has entered
the area around a feeding site and either rested or attempted to rest. If the site is not a \code{\link{FeedingSite}} or peri-domestic the mosquito always rests on vegetation.
\itemize{
\item i: rest on the inside wall of a structure
\item w: rest on the outside wall of a structure
\item v: rest on vegetation
\item r: reattempt without resting
\item l: leave the area
}
}
\details{
\itemize{
\item This method is bound to \code{MosquitoFemale$newSpot()}.
}
}
