% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Auxiliary.R
\docType{data}
\name{mbites_NULL}
\alias{mbites_NULL}
\title{M-BITES \code{NULL} Mosquito History}
\format{An object of class \code{list} of length 1.}
\usage{
mbites_NULL
}
\description{
This is used to pad the final entry when mosquitoes are being written to JSON by \code{\link[jsonlite]{toJSON}}.
}
\keyword{datasets}
