% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Complex-Bouts.R
\name{mbites_oneMosquito_MBITES}
\alias{mbites_oneMosquito_MBITES}
\title{M-BITES: Run Simulation for \code{\link{MosquitoFemale}}}
\usage{
mbites_oneMosquito_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm while alive and has not overrun time in enclosing \code{\link{Tile}}.
This method calls \code{\link{mbites_oneBout}} to simulate each life stage.
\itemize{
\item This method is bound to \code{MosquitoFemale$MBITES()}.
}
}
