% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_pEnergySurvival}
\alias{mbites_pEnergySurvival}
\title{MBITES-Generic: Probability of Death due to Energy Reserves for \code{\link{MosquitoFemale}}}
\usage{
mbites_pEnergySurvival()
}
\description{
Incremental mortality as a function of energy reserves given by \deqn{ \frac{e^{S.a\times energy}}{S.b+e^{S.a\times energy}} }
\itemize{
\item This method is bound to \code{MosquitoFemale$pEnergySurvival()}.
}
}
