% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Complex-Bouts.R
\name{mbites_Pop_MBITES}
\alias{mbites_Pop_MBITES}
\title{M-BITES: Run Simulation for \code{\link{MosquitoPopFemale}}}
\usage{
mbites_Pop_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm while alive and has not overrun time in enclosing \code{\link{MicroTile}}.
This method calls \code{\link{mbitesBRO_oneMosquito_MBITES}} one each living mosquito sequentially.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$MBITES()}.
}
}
