% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_queueSugarBout}
\alias{mbites_queueSugarBout}
\title{MBITES-Generic: Queue Sugar Bout due to Energy Reserves for \code{\link{MosquitoFemale}}}
\usage{
mbites_queueSugarBout()
}
\description{
Potentially queue a sugar bout \code{\link{mbites_boutS}} as a function of energy reserves.
\itemize{
\item This method is bound to \code{MosquitoFemale$queueSugarBout()}.
}
}
