% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_searchFail}
\alias{mbites_searchFail}
\title{M-BITES: Check State Transition for Resting Spot \code{MosquitoFemale}}
\usage{
mbites_searchFail()
}
\description{
Return \code{TRUE} if mosquito stays in "L" or "F" state.
}
\details{
\itemize{
\item This method is bound to \code{MosquitoFemale$searchFail()}.
}
}
