% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Energetics.R
\name{mbites_sugarEnergetics}
\alias{mbites_sugarEnergetics}
\title{MBITES-Generic: Sugar Energetics for \code{\link{MosquitoFemale}}}
\usage{
mbites_sugarEnergetics()
}
\description{
Handle energy dependent mortality and sugar bout queueing as function of current mosquito energy levels.
\itemize{
\item This method is bound to \code{MosquitoFemale$sugarEnergetics()}.
}
}
