% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Survival.R
\name{mbites_surviveResting}
\alias{mbites_surviveResting}
\title{MBITES-Generic: Resting Survival for \code{\link{MosquitoFemale}}}
\usage{
mbites_surviveResting()
}
\description{
Run generic resting survival probailities for bouts (launch to launch).
\itemize{
\item This method is bound to \code{MosquitoFemale$surviveResting()}.
}
}
