% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Bouts.R
\name{mbites_timingExponential}
\alias{mbites_timingExponential}
\title{M-BITES: Exponential Timing for \code{MosquitoFemale}}
\usage{
mbites_timingExponential()
}
\description{
Method for exponentially-distributed bout lengths (model mosquitoes as a Markov process).
\itemize{
\item This method is bound to \code{MosquitoFemale$timingExponential()}.
}
}
