% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Complex-Bouts.R
\name{mbites_timingGamma}
\alias{mbites_timingGamma}
\title{M-BITES: Gamma Timing for \code{MosquitoFemale}}
\usage{
mbites_timingGamma()
}
\description{
Method for Gamma-distributed bout lengths (model mosquitoes as a semi-Markov process).
\itemize{
\item This method is bound to \code{MosquitoFemale$timingExponential()}.
}
}
