% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Cohort.R
\name{mbitesBRO_cohort_simCohort}
\alias{mbitesBRO_cohort_simCohort}
\title{MBITES-BRO-Cohort: Run Simulation for \code{\link{MosquitoPopFemale}}}
\usage{
mbitesBRO_cohort_simCohort(N)
}
\arguments{
\item{N}{number of mosquitoes in cohort}
}
\description{
Run cohort simulation for population of mosquitoes and write bionomics to .json.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$simCohort()}.
}
}
