% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-BRO-Bouts.R
\name{mbitesBRO_timingExponential}
\alias{mbitesBRO_timingExponential}
\title{MBITES-BRO: Exponential Timing for \code{MosquitoFemale}}
\usage{
mbitesBRO_timingExponential()
}
\description{
Method for exponentially-distributed bout lengths (model mosquitoes as a Markov process).
\itemize{
\item This method is bound to \code{MosquitoFemale$timingExponential()}.
}
}
