% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Generic-HostEncounter.R
\name{mbitesGeneric_humanEncounter}
\alias{mbitesGeneric_humanEncounter}
\title{MBITES-BRO: Human Host Encounter for \code{\link{MosquitoFemale}}}
\usage{
mbitesGeneric_humanEncounter()
}
\description{
After calling \code{\link{mbitesGeneric_chooseHost}}, the mosquito encounters the human host and attempts to feed.
\itemize{
\item This method is bound to \code{MosquitoFemale$humanEncounter()}.
}
}
