% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_boutM}
\alias{mbitesMale_boutM}
\title{MBITES-Male: Male Mating Bout \code{MosquitoMale}}
\usage{
mbitesMale_boutM()
}
\description{
A male mosquito performs a mating bout. If successful (governed by parameter M_succeed), the mosquito enters a mating queue (see \code{\link[MASHcpp]{MatingQ}}).
\itemize{
\item This method is bound to \code{MosquitoMale$boutM()}.
}
}
