% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Survival.R
\name{mbitesMale_get_surviveFlightProb}
\alias{mbitesMale_get_surviveFlightProb}
\title{MBITES-Male: Get Baseline Survival Probability for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_get_surviveFlightProb()
}
\description{
Get baseline flight survival probability for \code{\link{mbitesGeneric_surviveFlight}}.
\itemize{
\item This method is bound to \code{MosquitoMale$get_surviveFlightProb()}.
}
}
