% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_get_WTS}
\alias{mbitesMale_get_WTS}
\title{MBITES-Male: Return Landing Spot Weights for \code{MosquitoMale}}
\usage{
mbitesMale_get_WTS()
}
\value{
vector of landing spot weights
}
\description{
Method for return landing spot weights based on behavioral state of mosquito.
\itemize{
\item This method is bound to \code{MosquitoMale$get_WTS()}.
}
}
