% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_landingSpot}
\alias{mbitesMale_landingSpot}
\title{MBITES-Male: Land After Flight \code{MosquitoMale}}
\usage{
mbitesMale_landingSpot()
}
\description{
Mosquito lands after a flight (choose a landing spot), which may cause various events.
This function always calls \code{\link{mbitesMale_newSpot}} and may call \code{\link{mbitesMale_enterHouse}}
Landing spots include:
\itemize{
\item i: 1 rest on the inside wall of a structure
\item w: 2 rest on the outside wall of a structure
\item v: 3 rest on vegetation
\item r: 4 reattempt without resting
\item l: 5 leave the area
}
}
\details{
\itemize{
\item This method is bound to \code{MosquitoMale$landingSpot()}.
}
}
