% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_newSpot}
\alias{mbitesMale_newSpot}
\title{MBITES-Male: Generate New Landing Spot for \code{MosquitoMale}}
\usage{
mbitesMale_newSpot()
}
\value{
character
}
\description{
Method for return a new landing spot based on behavioral state of mosquito and weights from \code{\link{mbitesMale_get_WTS}}.
If the site is not peri-domestic the mosquito always rests on vegetation.
\itemize{
\item i: rest on the inside wall of a structure
\item w: rest on the outside wall of a structure
\item v: rest on vegetation
\item r: reattempt without resting
\item l: leave the area
}
}
\details{
\itemize{
\item This method is bound to \code{MosquitoMale$newSpot()}.
}
}
