% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_Pop_MBITES}
\alias{mbitesMale_Pop_MBITES}
\title{MBITES-Male: Run Simulation for \code{\link{MosquitoPopMale}}}
\usage{
mbitesMale_Pop_MBITES()
}
\description{
Run the M-BITES life cycle simulation algorithm while alive and has not overrun time in enclosing \code{\link{MicroTile}}.
This method calls \code{\link{mbitesMale_oneMosquito_MBITES}} one each living mosquito sequentially.
Prior to running simulation, this function calls \code{\link{clear_MatingQ_Landscape}} to clear mating queues.
\itemize{
\item This method is bound to \code{MosquitoPopMale$MBITES()}.
}
}
