% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Energetics.R
\name{mbitesMale_pSugarBout}
\alias{mbitesMale_pSugarBout}
\title{MBITES-Male: Probability to Queue Sugar Bout due to Energy Reserves for \code{\link{MosquitoMale}}}
\usage{
mbitesMale_pSugarBout()
}
\description{
Calculate sugar bout probability as a function of energy reserves.
\itemize{
\item This method is bound to \code{MosquitoMale$pSugarBout()}.
}
}
