% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_restingSpot}
\alias{mbitesMale_restingSpot}
\title{MBITES-Male: Land After Flight \code{MosquitoMale}}
\usage{
mbitesMale_restingSpot()
}
\description{
Mosquito lands after a flight (choose a landing spot), which may cause various events.
This function always calls \code{\link{mbitesMale_newSpot}} and may call \code{\link{mbitesMale_enterHouse}}
Landing spots include:
\itemize{
\item i: rest on the inside wall of a structure
\item w: rest on the outside wall of a structure
\item v: rest on vegetation
\item r: reattempt without resting
\item l: leave the area
}
}
\details{
\itemize{
\item This method is bound to \code{MosquitoMale$restingSpot()}.
}
}
