% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_timingExponential}
\alias{mbitesMale_timingExponential}
\title{MBITES-Male: Exponential Timing for \code{MosquitoMale}}
\usage{
mbitesMale_timingExponential()
}
\description{
Method for exponentially-distributed bout lengths (model mosquitoes as a Markov process).
\itemize{
\item This method is bound to \code{MosquitoMale$timingExponential()}.
}
}
