% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Bouts.R
\name{mbitesMale_timingGamma}
\alias{mbitesMale_timingGamma}
\title{MBITES-Male: Gamma Timing for \code{MosquitoMale}}
\usage{
mbitesMale_timingGamma()
}
\description{
Method for Gamma-distributed bout lengths (model mosquitoes as a semi-Markov process).
\itemize{
\item This method is bound to \code{MosquitoMale$timingExponential()}.
}
}
