% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Utility.R
\name{MicroKernel_FitExpCDF}
\alias{MicroKernel_FitExpCDF}
\title{MICRO Kernels: Fit Exponential Lambda}
\usage{
MicroKernel_FitExpCDF(d, q, up = 1)
}
\arguments{
\item{d}{numeric distance}

\item{q}{numeric quantile (must be between 0 and 1)}

\item{up}{numeric upper limit on rate parameter (can be set to 1/(max distance anything could happen at))}
}
\description{
Fit an exponential distribution by quantile such that 'quantile' amount of activity
occurs at less than or equal to input 'distance' d. Returns fitted rate parameter fitted from \code{\link[base]{optimise}}.
}
