% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Methods.R
\name{MicroKernel_moveMe_BROM}
\alias{MicroKernel_moveMe_BROM}
\title{MICRO Search Kernels: \code{\link{MosquitoFemale}} Movement Function for Full M-BITES Lifecycle Model}
\usage{
MicroKernel_moveMe_BROM()
}
\description{
Move one mosquito based on site and next behavioral state. This method is bound to \code{MosquitoFemale$moveMe()}
}
