% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Methods.R
\name{MicroKernel_SampleMvOb_MosquitoPop}
\alias{MicroKernel_SampleMvOb_MosquitoPop}
\title{MICRO Search Kernels: \code{\link{MicroMosquito}} Sample MvOb}
\usage{
MicroKernel_SampleMvOb_MosquitoPop(ixS, state, inPointSet)
}
\arguments{
\item{ixS:}{current site index \code{ix} of this mosquito}

\item{state:}{current behavioral state of this mosquito}

\item{inPointSet:}{current point set this mosquito is in}
}
\value{
new index
}
\description{
This method is a helper for \code{\link{MicroKernel_moveMe}} and samples the appropriate MvOb in MvAll stored in an enclosing \code{\link{Tile}} object.
\itemize{
\item This function is bound to \code{MosquitoPopFemale$SampleMove} and \code{MosquitoPopMale$SampleMove}
}
}
