% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Utility.R
\name{MicroKernelPlotOne_utility}
\alias{MicroKernelPlotOne_utility}
\title{MICRO Search Kernels: Plot CDF and PDF for One Starting Site}
\usage{
MicroKernelPlotOne_utility(startXY, endPts, sigma = 3, eps = 0.1,
  beta = 0)
}
\arguments{
\item{endPts}{list of end points and weights (output of \code{\link{get_PointSet_Landscape}})}

\item{sigma}{a param}

\item{eps}{a param}

\item{beta}{a param}
}
\description{
for manuscript. write me!
}
