% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{push_pop_MosquitoPopMale}
\alias{push_pop_MosquitoPopMale}
\title{Push pop}
\usage{
push_pop_MosquitoPopMale(N, tEmerge, genotype, locNow)
}
\arguments{
\item{N}{integer; number of emerging mosquitoes}

\item{tEmerge}{integer; day of emergence}

\item{genotype}{integer; genotype of emerging mosquitoes}

\item{locNow}{integer; site of emergence}
}
\description{
Push new male mosquitoes into \code{\link{MosquitoPopMale}}.
\itemize{
\item This method is bound to \code{MosquitoPopMale$push_pop}
}
}
