% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Simulation.R
\name{reset_HumanPop_Tile}
\alias{reset_HumanPop_Tile}
\title{MICRO \code{\link{MicroTile}}: Re-initialize \code{\link[MASHmacro]{HumanPop}}}
\usage{
reset_HumanPop_Tile(HumanPop_PAR)
}
\arguments{
\item{HumanPop_PAR}{from \code{\link[MASHmacro]{HumanPop.Parameters}}}
}
\description{
write me! used to reset humanpop between simulation runs
}
