% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Setup.R
\name{SEARCH.Kernel.Setup}
\alias{SEARCH.Kernel.Setup}
\title{MICRO Search Kernels: Initialize Additional Methods & Fields in \code{\link{MosquitoPopFemale}} and \code{\link{MosquitoFemale}}}
\usage{
SEARCH.Kernel.Setup(MBITES = "BRO", overwrite = TRUE)
}
\arguments{
\item{MBITES}{what M-BITES module to run?
\itemize{
\item BRO: Blood Feeding, Resting, Oviposition module (also including BRO with SUGAR and/or MATE)
\item FULL: Full life cycle M-BITES module (Blood Feeding Search & Attempt, Resting, Oviposition Search & Attempt, Mating, Sugar Feeding)
}}
}
\description{
Initialize MICRO Search Kernels module of mosquito search behavior.
}
