% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods-VectorControl.R
\name{set_aerialSpray}
\alias{set_aerialSpray}
\title{Set aerialSpray}
\usage{
set_aerialSpray(aerialSprayDefinition)
}
\arguments{
\item{aerialSpray}{Object}
}
\description{
Change the definition of the Aerial Spray
}
