% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_AquaSites_Landscape}
\alias{set_AquaSites_Landscape}
\title{Set Aquatic Habitat Sites}
\usage{
set_AquaSites_Landscape(AquaSites, ixS = NULL)
}
\arguments{
\item{AquaSites}{input may be a single \code{\link{AquaticSite}} or list of sites}

\item{ixS}{integer index; if \code{NULL} set entire list of sites}
}
\description{
Set a single \code{\link{AquaticSite}} or all sites.
\itemize{
\item This method is bound to \code{Landscape$set_AquaSites}
}
}
