% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_d_E_EL4P}
\alias{set_d_E_EL4P}
\title{EL4P: Set d_E}
\usage{
set_d_E_EL4P(d_E)
}
\arguments{
\item{d_E}{numeric}
}
\description{
Set d_E
\itemize{
\item This method is bound to \code{EL4P$set_d_E}
}
}
