% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_d_L_EL4P}
\alias{set_d_L_EL4P}
\title{EL4P: Set d_L}
\usage{
set_d_L_EL4P(d_L)
}
\arguments{
\item{d_L}{numeric}
}
\description{
Set d_L
\itemize{
\item This method is bound to \code{EL4P$set_d_L}
}
}
