% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_d_P_EL4P}
\alias{set_d_P_EL4P}
\title{EL4P: Set d_P}
\usage{
set_d_P_EL4P(d_P)
}
\arguments{
\item{d_P}{numeric}
}
\description{
Set d_P
\itemize{
\item This method is bound to \code{EL4P$set_d_P}
}
}
