% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_dE_EL4P}
\alias{set_dE_EL4P}
\title{EL4P: Set dE}
\usage{
set_dE_EL4P(dE)
}
\arguments{
\item{dE}{numeric}
}
\description{
Set dE
\itemize{
\item This method is bound to \code{EL4P$set_dE}
}
}
