% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_dEgg_EL4P}
\alias{set_dEgg_EL4P}
\title{EL4P: Set dEgg}
\usage{
set_dEgg_EL4P(dEgg)
}
\arguments{
\item{dEgg}{numeric}
}
\description{
Set dEgg
\itemize{
\item This method is bound to \code{EL4P$set_dEgg}
}
}
