% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_dL_EL4P}
\alias{set_dL_EL4P}
\title{EL4P: Set dL}
\usage{
set_dL_EL4P(dL)
}
\arguments{
\item{dL}{numeric}
}
\description{
Set dL
\itemize{
\item This method is bound to \code{EL4P$set_dL}
}
}
