% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_dP_EL4P}
\alias{set_dP_EL4P}
\title{EL4P: Set dP}
\usage{
set_dP_EL4P(dP)
}
\arguments{
\item{dP}{numeric}
}
\description{
Set dP
\itemize{
\item This method is bound to \code{EL4P$set_dP}
}
}
