% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_dPupae_EL4P}
\alias{set_dPupae_EL4P}
\title{EL4P: Set dPupae}
\usage{
set_dPupae_EL4P(dPupae)
}
\arguments{
\item{dPupae}{numeric}
}
\description{
Set dPupae
\itemize{
\item This method is bound to \code{EL4P$set_dPupae}
}
}
