% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_enterP_FeedingSite}
\alias{set_enterP_FeedingSite}
\title{Set enterP}
\usage{
set_enterP_FeedingSite(enterP)
}
\arguments{
\item{enterP}{numeric}
}
\description{
Set landing hazard associated with indoor wall at this site \code{enterP}
\itemize{
\item This method is bound to \code{FeedingSite$set_enterP}
}
}
