% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-HumanPop-Setup.R
\name{set_FemalePopPointer_HumanPop}
\alias{set_FemalePopPointer_HumanPop}
\title{Set \code{\link{MosquitoPopFemale}} Pointer}
\usage{
set_FemalePopPointer_HumanPop(FemalePopPointer)
}
\arguments{
\item{FemalePopPointer}{an environment}
}
\description{
Set pointer to the female mosquito population on this landscape.
\itemize{
\item This method is bound to \code{HumanPop$set_FemalePopPointer}
}
}
