% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Landscape-Methods.R
\name{set_HumansPointer_Landscape}
\alias{set_HumansPointer_Landscape}
\title{Set \code{\link[MASHmacro]{HumanPop}} Pointer}
\usage{
set_HumansPointer_Landscape(HumansPointer)
}
\arguments{
\item{HumansPointer}{an environment}
}
\description{
Set pointer to \code{HumanPop$self}.
\itemize{
\item This method is bound to \code{Landscape$set_HumansPointer}
}
}
