% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_inPointSet_Mosquito}
\alias{set_inPointSet_Mosquito}
\title{Set inPointSet}
\usage{
set_inPointSet_Mosquito(inPointSet)
}
\arguments{
\item{inPointSet}{character}
}
\description{
Set mosquito point set (\code{character}). With \code{ix}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_inPointSet} and \code{MosquitoMale$set_inPointSet}
}
}
