% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_ix_Mosquito}
\alias{set_ix_Mosquito}
\title{Set ix}
\usage{
set_ix_Mosquito(ix)
}
\arguments{
\item{ix}{integer}
}
\description{
Set mosquito point index (\code{integer}). With \code{inPointSet}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_ix} and \code{MosquitoMale$set_ix}
}
}
