% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_L1_EL4P}
\alias{set_L1_EL4P}
\title{EL4P: Set L1}
\usage{
set_L1_EL4P(L1)
}
\arguments{
\item{L1}{numeric}
}
\description{
Set L1
\itemize{
\item This method is bound to \code{EL4P$set_L1}
}
}
