% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-Emerge-Methods.R
\name{set_lambda_Emerge}
\alias{set_lambda_Emerge}
\title{Aquatic Ecology \code{\link{AquaticSite}} Method: Set Lambda}
\usage{
set_lambda_Emerge(lambda)
}
\arguments{
\item{lambda}{numeric vector}
}
\description{
Set lambda.
This method is bound to \code{AquaticSite$set_lambda()}.
}
