% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-PeriDomestic.R
\name{set_lambda_periDomestic_AquaticSite}
\alias{set_lambda_periDomestic_AquaticSite}
\title{Set lambda}
\usage{
set_lambda_periDomestic_AquaticSite(lambda)
}
\arguments{
\item{lambda}{numeric(365)}
}
\description{
Set vector of lambda of mean daily emergence at this site
\itemize{
\item This method is bound to \code{periDomestic_AquaticSite$set_lambda}
}
}
