% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_locNext_Mosquito}
\alias{set_locNext_Mosquito}
\title{Set locNext}
\usage{
set_locNext_Mosquito(loc)
}
\arguments{
\item{loc}{character}
}
\description{
Set current mosquito point set (\code{character}). With \code{pSetNext}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_locNext} and \code{MosquitoMale$set_locNext}
}
}
