% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_locOld_Mosquito}
\alias{set_locOld_Mosquito}
\title{Set locOld}
\usage{
set_locOld_Mosquito(loc)
}
\arguments{
\item{loc}{character}
}
\description{
Set current mosquito point set (\code{character}). With \code{pSetOld}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_locOld} and \code{MosquitoMale$set_locOld}
}
}
