% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_MaleCSVCon_Tile}
\alias{set_MaleCSVCon_Tile}
\title{Set Male CSV Connection}
\usage{
set_MaleCSVCon_Tile(MaleCSVCon)
}
\arguments{
\item{MaleCSVCon}{object of class \code{\link[base]{connection}}, usually opened with \code{file}}
}
\description{
Set the text connection used to write population count data to CSV.
\itemize{
\item This method is bound to \code{Tile$set_MaleCSVCon}
}
}
