% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{set_MaleHistoryCon_Tile}
\alias{set_MaleHistoryCon_Tile}
\title{Set Male JSON Connection}
\usage{
set_MaleHistoryCon_Tile(MaleHistoryCon)
}
\arguments{
\item{MaleHistoryCon}{object of class \code{\link[base]{connection}}, usually opened with \code{file}}
}
\description{
Set the text connection used to write male histories to JSON.
\itemize{
\item This method is bound to \code{Tile$set_MaleHistoryCon}
}
}
