% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-MosquitoPop-Methods.R
\name{set_MBITES_PAR_MosquitoPop}
\alias{set_MBITES_PAR_MosquitoPop}
\title{Set MBITES_PAR}
\usage{
set_MBITES_PAR_MosquitoPop(MBITES_PAR)
}
\description{
Set M-BITES parameter list.
\itemize{
\item This method is bound to \code{MosquitoPopFemale$set_MBITES_PAR} and \code{MosquitoPopFemale$set_MBITES_PAR}
}
}
