% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_muE_EL4P}
\alias{set_muE_EL4P}
\title{EL4P: Set muE}
\usage{
set_muE_EL4P(muE)
}
\arguments{
\item{muE}{numeric}
}
\description{
Set muE
\itemize{
\item This method is bound to \code{EL4P$set_muE}
}
}
