% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_muEgg_EL4P}
\alias{set_muEgg_EL4P}
\title{EL4P: Set muEgg}
\usage{
set_muEgg_EL4P(muEgg)
}
\arguments{
\item{muEgg}{numeric}
}
\description{
Set muEgg
\itemize{
\item This method is bound to \code{EL4P$set_muEgg}
}
}
