% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_muL_EL4P}
\alias{set_muL_EL4P}
\title{EL4P: Set muL}
\usage{
set_muL_EL4P(muL)
}
\arguments{
\item{muL}{numeric}
}
\description{
Set muL
\itemize{
\item This method is bound to \code{EL4P$set_muL}
}
}
