% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_muLarva_EL4P}
\alias{set_muLarva_EL4P}
\title{EL4P: Set muLarva}
\usage{
set_muLarva_EL4P(muLarva)
}
\arguments{
\item{muLarva}{numeric}
}
\description{
Set muLarva
\itemize{
\item This method is bound to \code{EL4P$set_muLarva}
}
}
