% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AQUA-EL4P-Methods.R
\name{set_muPupae_EL4P}
\alias{set_muPupae_EL4P}
\title{EL4P: Set muPupae}
\usage{
set_muPupae_EL4P(muPupae)
}
\arguments{
\item{muPupae}{numeric}
}
\description{
Set muPupae
\itemize{
\item This method is bound to \code{EL4P$set_muPupae}
}
}
