% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Site-Methods.R
\name{set_numGenotypes_AquaticSite}
\alias{set_numGenotypes_AquaticSite}
\title{Set numGenotypes}
\usage{
set_numGenotypes_AquaticSite(numGenotypes)
}
\arguments{
\item{numGenotypes}{numeric(365)}
}
\description{
Set number of genotypes.
\itemize{
\item This method is bound to \code{AquaticSite$set_numGenotypes}
}
}
