% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_pSetNext_Mosquito}
\alias{set_pSetNext_Mosquito}
\title{Set pSetNext}
\usage{
set_pSetNext_Mosquito(pSet)
}
\arguments{
\item{pSet}{character}
}
\description{
Set current mosquito point set (\code{character}). With \code{locNow}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_pSetNext} and \code{MosquitoMale$set_pSetNext}
}
}
