% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{set_pSetNow_Mosquito}
\alias{set_pSetNow_Mosquito}
\title{Set pSetNow}
\usage{
set_pSetNow_Mosquito(pSet)
}
\arguments{
\item{pSet}{character}
}
\description{
Set current mosquito point set (\code{character}). With \code{locNow}, the mosquito spatial location is resolved.
\itemize{
\item This method is bound to \code{MosquitoFemale$set_pSetNow} and \code{MosquitoMale$set_pSetNow}
}
}
