% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HUMAN-ActivitySpace-Methods.R
\name{sim_ActivitySpace_HumanPop}
\alias{sim_ActivitySpace_HumanPop}
\title{MICRO \code{\link{HumanPop}} Method: Simulate Activity Space}
\usage{
sim_ActivitySpace_HumanPop()
}
\description{
This function is bound to \code{HumanPop$sim_ActivitySpace()}
}
