% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Simulation.R
\name{simMICRO_oneRun}
\alias{simMICRO_oneRun}
\title{MICRO \code{\link{MicroTile}}: One Simulation Run}
\usage{
simMICRO_oneRun(tMax, PfPAR, verbose = FALSE, trackPop = FALSE)
}
\arguments{
\item{tMax}{length of simulation run}

\item{PfPAR}{parameters passed to \code{HumanPop$initialize_Pathogens}}

\item{verbose}{print information}

\item{trackPop}{write mosquito population counts to CSV file}
}
\description{
write me
\itemize{
\item This method is bound to \code{Tile$simMICRO_oneRun}
}
}
