% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Simulation.R
\name{simMICRO_oneStep}
\alias{simMICRO_oneStep}
\title{MICRO \code{\link{MicroTile}}: Run Simulation one Time Step}
\usage{
simMICRO_oneStep(trackPop = FALSE)
}
\arguments{
\item{timeStep}{the time step of the model}

\item{clearInterval}{interval to clear population and track mosquito history (see \code{\link{clear_pop}})}

\item{popTrack}{log daily counts of mosquitoes in life stages or not (written to MicroTile$directory)}

\item{historyTrack}{output mosquito histories to JSON via \code{\link{MosquitoPopFemale_clear_pop}} every clearInterval days or not (written to MicroTile$directory)}
}
\description{
Run MICRO simulation for one time step, the length of which defines the temporal window for indifference to contingent events.
}
