% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Class.R
\docType{class}
\name{Tile}
\alias{Tile}
\title{Tile Class Definition}
\format{An \code{\link{R6Class}} generator object}
\usage{
Tile
}
\description{
im a class!
}
\section{\strong{Constructor}}{

\itemize{
\item argument: im an agument!
}
}

\section{\strong{Methods}}{

\itemize{
\item method: im a method!
}
}

\section{\strong{Fields}}{

\itemize{
\item field: im a field!
}
}

\keyword{R6}
\keyword{class}
