% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MBITES-Male-Survival.R
\name{mbitesMale_rTatterSize}
\alias{mbitesMale_rTatterSize}
\title{MBITES-Generic: Wing Tattering for \code{\link{MosquitoFemale}}}
\usage{
mbitesMale_rTatterSize()
}
\description{
Draw from a zero-inflated Beta distribution for additive wing damage from tattering.
Wing damage is given by \deqn{ \left\{\begin{matrix}
x=0; P(ttsz.p.m)
\\
x\sim Beta(ttsz.a.m,ttsz.b.m); P(1-ttsz.p.m)
\end{matrix}\right. }
\itemize{
\item This method is bound to \code{MosquitoMale$rTatterSize()}.
}
}
