% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEARCH-Kernel-Setup.R
\name{SEARCH.Kernel.Setup}
\alias{SEARCH.Kernel.Setup}
\title{MICRO Search Kernels: Initialize Additional Methods & Fields in \code{\link{MosquitoPopFemale}} and \code{\link{MosquitoFemale}}}
\usage{
SEARCH.Kernel.Setup(MBITES = "BRO", overwrite = TRUE)
}
\arguments{
\item{MBITES}{what M-BITES module to run?
\itemize{
\item BRO: Blood Feeding, Resting, Oviposition module
\item BROM: Blood Feeding, Resting, Oviposition, Mating module
\item FULL: Full life cycle M-BITES module
}}
}
\description{
Initialize MICRO Search Kernels module of mosquito search behavior.
}
