% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tile-Methods.R
\name{simMICRO_oneRun}
\alias{simMICRO_oneRun}
\title{MICRO \code{\link{MicroTile}}: One Simulation Run}
\usage{
simMICRO_oneRun(tMax, runID, verbose = FALSE, timeStep = 1,
  clearInterval = 10, popTrack = FALSE, historyTrack = FALSE)
}
\arguments{
\item{tMax}{length of simulation run}

\item{runID}{unique simulation run identifier}

\item{verbose}{print information}

\item{timeStep}{the time step of the model}

\item{clearInterval}{interval to clear population and track mosquito history (see \code{\link{clear_pop}})}

\item{popTrack}{log daily counts of mosquitoes in life stages or not (written to MicroTile$directory)}

\item{historyTrack}{output mosquito histories to JSON via \code{\link{MicroMosquitoPopFemale_clear_pop}} every clearInterval days or not (written to MicroTile$directory)}
}
\description{
write me
}
