% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MOSQUITO-Mosquito-Methods.R
\name{writeAndDelete_MosquitoFemale}
\alias{writeAndDelete_MosquitoFemale}
\title{Export JSON History and Remove Self}
\usage{
writeAndDelete_MosquitoFemale(conHist, conPath)
}
\arguments{
\item{conHist}{a \code{\link[base]{connection}} to write mosquito history to (the correct directory can be found with \code{\link{get_MosquitoDirectory_Tile}})}

\item{conPath}{a \code{\link[base]{connection}} to write mosquito-stage pathogen history to (the correct directory can be found with \code{\link{get_MosquitoDirectory_Tile}})}
}
\description{
\itemize{
\item This method is bound to \code{MosquitoFemale$writeAndDelete}
}
}
